Reval n Hits update
Now the opponent (AI) also presses the buttons. We've also added button press effects that show you exactly how you hit the button to the beat of the music. Because of this, a bug appeared, due to which, under an unknown set of circumstances, sometimes when the button is pressed in time, the inscription "miss" still appears.